권한 없는 사용자가 수정, 삭제하는 버그 수정 (#39)
* refactor: RestApi형식으로 리펙토링

* refactor: rest형식으로 리펙토링

* fix: SiteUser -> Member

* fix: 단일 아티클 조회를 실패하는 오류를 수정

* delete: db파일 삭제

* refactor: 각각의 메서드별로 나눔

* fix: Setter로 변경

* fix: 인증작업을 제대로 하지 않는 오류 수정

* fix: 권한 없는 사용자가 수정, 삭제하는 버그 수정

* refactor: 필요없는 코드 삭제

* feat: ExceptioHandler 추가

* fix: 데이터 검증작업을 하지않는 오류 수정

* fix: ""빈 문자열 받아내는 오류 수정

+++ src/main/java/com/example/techit7/article/service/ArticleServiceImpl.java
@@ -8,9 +8,12 @@ import com.example.techit7.article.entity.Article;
 import com.example.techit7.article.repository.ArticleRepository;
 import com.example.techit7.global.response.GlobalResponse;
 import com.example.techit7.user.entity.Member;
+import com.example.techit7.user.service.MemberRestServiceImpl;
+import io.jsonwebtoken.security.Jwks.OP;
 import jakarta.persistence.EntityNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -85,11 +88,14 @@ public class ArticleServiceImpl implements ArticleService {
      */
     @Override
     @Transactional
-    public void updateArticleById(Long id, ArticleRequestDto articleRequestDto) {
+    public void updateArticleById(Long id, ArticleRequestDto articleRequestDto, String username) {
         Optional<Article> article = articleRepository.findById(id);
 
         if (article.isEmpty()) {
             throw new EntityNotFoundException(ENTITY_NOT_FOUND + id);
+        }
+        if (!article.get().getAuthor().getUsername().equals(username)) {
+            throw new IllegalArgumentException("권한이 없는 유저입니다.");
         }
 
         article.get().modifyArticle(articleRequestDto);
@@ -101,12 +107,17 @@ public class ArticleServiceImpl implements ArticleService {
      */
     @Override
     @Transactional
-    public void deleteArticleById(Long id) {
+    public void deleteArticleById(Long id, String username) {
+
         Optional<Article> article = articleRepository.findById(id);
 
         if (article.isEmpty()) {
             throw new EntityNotFoundException(ENTITY_NOT_FOUND + id);
         }
+        if (!article.get().getAuthor().getUsername().equals(username)) {
+            throw new IllegalArgumentException("권한이 없는 유저입니다.");
+        }
+
         articleRepository.delete(article.get());
     }
